+++ pkg/hookbot/auth.go
@@ -1,4 +1,4 @@
-package main
+package hookbot
 
 import (
 	"crypto/hmac"
