+++ pkg/hookbot/util.go
@@ -1,4 +1,4 @@
-package main
+package hookbot
 
 import (
 	"log"
